Create Contact component
Create separate Contact component to represent a single entry
in the list of contacts.  This will be called from the
ListContacts component and will take in a single contact to
return a list item representing that contact.

Each list item displays the contact avatar, name, social media
handle, and a button which will eventually be connected to allow
you to delete that contact.